+++ CHANGELOG.md
@@ -1,15 +1,25 @@
 # Change Log
+
+## v0.1.0
+- Added mdBook functionality to serve static website with hotkey
+- Added search notes functionality with hotkey
+- Same concept as Vimwiki with a dedicated folder for notes
+- Fixed async issues, will correctly wait for install and running kernel before executing cell now
+- Using tmp folder to store code
+- Add hotkey to preview main.rs from generated code
+- Added language provider (just getting `println` to display)
+
 ## v0.0.9
-Bugfix with output spacing
+- Bugfix with output spacing
 
 ## v0.0.8
-Now sends the workspace folder name to `rustkernel`
+- Now sends the workspace folder name to `rustkernel`
 
 ## v0.0.7
-Fix bug with editor adding new lines
+- Fix bug with editor adding new lines
 
 ## v0.0.4 
-All dependencies removed
+- All dependencies removed
 
 ## v0.0.3 
-mvp
+- mvp